Split mine status model class file
