+++ python-backend/app/api/mines/mine/resources/mine.py
@@ -5,7 +5,8 @@ from flask import request
 from flask_restplus import Resource, reqparse, inputs
 from sqlalchemy_filters import apply_sort, apply_pagination
 
-from ...status.models.status import MineStatus, MineStatusXref
+from ...status.models.mine_status import MineStatus
+from ...status.models.mine_status_xref import MineStatusXref
 from ..models.mine_identity import MineIdentity
 from ..models.mine_detail import MineDetail
 from ..models.mineral_tenure_xref import MineralTenureXref
