+++ python-backend/tests/conftest.py
@@ -8,7 +8,9 @@ from app.extensions import db, jwt as _jwt
 from app.api.mines.mine.models.mine_identity import MineIdentity
 from app.api.mines.mine.models.mine_detail import MineDetail
 from app.api.mines.mine.models.mineral_tenure_xref import MineralTenureXref
-from app.api.mines.status.models.status import MineOperationStatusCode, MineOperationStatusReasonCode, MineOperationStatusSubReasonCode
+from app.api.mines.status.models.mine_operation_status_code import MineOperationStatusCode
+from app.api.mines.status.models.mine_operation_status_reason_code import MineOperationStatusReasonCode
+from app.api.mines.status.models.mine_operation_status_sub_reason_code import MineOperationStatusSubReasonCode
 from app.api.parties.party.models.party import Party, MgrAppointment, PartyTypeCode
 from app.api.mines.location.models.mine_location import MineLocation
 from app.api.permits.permit.models.permit import Permit, PermitStatusCode
